# license-server/openapi.yaml
openapi: 3.0.3
info:
  title: SafeSoundArena License Server API
  description: |
    API לשרת רישיונות SafeSoundArena המאפשר ניהול ואימות רישיונות.
    
    שרת זה מספק יכולות ליצירה, עדכון, מחיקה ואימות של רישיונות עבור מוצרי SafeSoundArena.
  version: 1.0.0
  contact:
    name: צוות SafeSoundArena
    email: support@safesoundarena.com
    url: https://safesoundarena.com
  license:
    name: ראה LICENSE.md בתיקיית השורש של הפרויקט
servers:
  - url: https://license.safesoundarena.com
    description: שרת ייצור
  - url: http://localhost:3010
    description: שרת פיתוח מקומי
...
paths:
...
  /admin/stats:
    get:
      tags:
        - admin
      summary: קבלת סטטיסטיקות
      description: מחזיר סטטיסטיקות על הרישיונות במערכת
      operationId: getStats
      security:
        - BearerAuth: []
      responses:
        '200':
          description: סטטיסטיקות רישיונות
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse'
        '401':
          description: לא מורשה
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: אין הרשאות מספיקות
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StatsResponse:
      type: object
      required:
        - totalLicenses
        - activeLicenses
        - licensesByType
        - verificationCount
        - lastUpdated
      properties:
        totalLicenses:
          type: integer
          example: 100
        activeLicenses:
          type: integer
          example: 85
        expiredLicenses:
          type: integer
          example: 15
        licensesByType:
          type: object
          additionalProperties:
            type: integer
          example:
            premium: 50
            basic: 30
            enterprise: 20
        verificationCount:
          type: integer
          example: 1250
        lastUpdated:
          type: string
          format: date-time
          example: '2023-06-15T08:30:00Z'
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Invalid license key